requirements.txt
feedparser
newspaper3k
transformers
torch
sqlalchemy
pymysql
apscheduler
fastapi
uvicorn
jinja2
aiofiles
python-multipart
